SM-6 split logib endpoint to new definiton to follow REST principle
